+++ packages/epub-parser/src/model/ParseContext.js
@@ -3,6 +3,7 @@ import Book from './Book';
 class ParseContext {
   constructor() {
     this.options = undefined;
+    this.entries = undefined;
     this.opfPath = undefined;
     this.basePath = undefined;
     this.foundCover = false;