+++ packages/pdf-parser/src/model/Book.js
@@ -1,0 +1,17 @@
+import Item from './Item';
+
+class Book {
+  constructor(rawBook = {}) {
+    this.items = (rawBook.items || []).map(rawObj => new Item(rawObj));
+    [this.cover] = this.items;
+    Object.freeze(this);
+  }
+
+  toRaw() {
+    return {
+      items: this.items.map(item => item.toRaw()),
+    };
+  }
+}
+
+export default Book;