+++ packages/pdf-parser/src/model/ReadContext.js
@@ -1,0 +1,9 @@
+class ReadContext {
+  constructor() {
+    this.items = undefined;
+    this.entries = undefined;
+    this.options = undefined;
+  }
+}
+
+export default ReadContext;