+++ packages/pdf-parser/test/model/ParseContext.spec.js
@@ -1,0 +1,14 @@
+import { should } from 'chai';
+
+import ParseContext from '../../src/model/ParseContext';
+
+should(); // Initialize should
+
+describe('Model - ParseContext', () => {
+  it('Initialize test', () => {
+    const item = new ParseContext();
+    Object.keys(item).forEach((key) => {
+      Object.getOwnPropertyDescriptor(item, key).writable.should.be.true;
+    });
+  });
+});